Revamp linking.
Redesign how mustang ensures its libraries are linked in, to be somewhat
more robust in the face of various kinds of optimizations. Use explicit symbols
to pull in individual modules. This also makes it easier to split more c-scape
functionality into separate modules.

 - Instead of having users declare `extern crate mustang`, have them invoke
   a macro, `mustang::can_compile_this!();`.

 - Don't declare c-scape functions as `pub`. Most code should use the
   `libc` crate declarations; `c-scape` uses `extern "C"` to intercept
   those at link time.

 - Put `extern "C"` functions in a named section; this overrides
   -ffunction-sections and puts all the functions in the same section,
   so that the linker doesn't DCE individual functions before they have a
   chance to intercept libc calls.

 - Similarly, add `#[inline(never)]` so that these functions aren't inlined,
   which can result in symbols being removed if an optimizer believes it
   has seen all of their uses.

This helps with #7, though is not the only thing needed.