Go back to explicitly defining `memcpy` etc.
Referencing "memcpy" and other symbols from `__mustang_c_scape__c`
didn't turn out to be sufficient to ensure they'd be linked in from
compiler-builtins, and not from the host libc.

Switch back to explicitly defining `memcpy` etc., though still use
compiler-builtins to do the work. This requires us to use
`#[feature(rustc_private)]` for now, but fortunately the things we're
depending on should be relatively stable in practice.

Also add a test for this.