Remove the initialize-c-runtime feature.
It doesn't look like initializing the C runtime is going to be useful;
we don't want the C runtime calling its own `malloc`, for example, because
we replace `malloc` for user code. So the new path forward is, mustang
is the C runtime. It doesn't have `printf` or `qsort` or `strcat` or
`catopen` or many other things; if these turn out to be needed, we'll
either have to implement them, or figure out a way to not need them.