Use `dlmalloc` as the default allocator.
The `dlmalloc` crate is a port of the dlmalloc library to Rust.

`wee_alloc` is specialized for small code size. Retain it as an option,
but use `dmalloc` by default, as it's a general-purpose allocator.

Also, move the `#[global_allocator]` code out of c-scape and into
mustang, as the roles for these crates are emerging. c-scape is focused
on providing C ABI compatibility, while mustang is focused on pulling
together all the pieces to support the *-mustang-* targets.

Fixes #28.