Use `FnOnce` functions instead of `unsafe extern "C"` functions.
Change origin's APIs which take callsbacks to take `Box<dyn FnOnce()>`
instead of `unsafe extern "C"` functions with `*mut c_void` callbacks,
so that `at_exit` and similar functions can be safe.

This also simplifies the `clone` assembly code sequences, as they no
longer need to pass a separate `arg` argument through, which previously
meant storing it on the child stack on some architectures.

It does, however, mean we double-`Box`, once to box up the `dyn Any`
object and another box so that we can get a single non-fat pointer to
it. If Rust ever gains a way to manipulate the raw parts of a fat
pointer and pass them through assembly code we might be able to avoid
this outer `Box`, but at present, double-`Box`ing seems to be the
recommended way.