Use a futex-based lock implementations. (#113)
Use the new futex-based lock implementations from std to provide
`Mutex`, `RwLock`, and `Condvar`. We unfortunately need to copy the
code for now, to break a dependency cycle with std, given that Mustang
is positioned underneath std.

This replaces the use of parking_lot, and avoids a lot of tricky
challenges related to parking_lot's dynamic allocations.