Use cargo features rather than `target_vendor` in c-scape and c-gull
Add cargo features "take-charge" and "coexist-with-libc" to c-scape
and c-gull, and use those features instead of testing for
`target_vendor = "mustang"`.

This is similar to how origin was recently converted to use cargo
features instead of `target_vendor = "mustang"` and will enable
c-scape and c-gull to be used in more ways.