Move c-scape and c-gull into their own repo.
Similar to origin, move c-scape and c-gull into their own repo, which is now
[here].

With this change, there's very little code left in the mustang repo here; it's
just the custom target, some minimal glue code, and all the tests and examples.

This also renames the "threads" feature to "thread" following origin and now
also c-gull and c-scape.

[here]: https://github.com/sunfishcode/c-ward