Move the mustang crate to the top level of the repo.
Previously this repo had a mustang-examples crate at the top level
with examples and tests, and a mustang directory containing the actual
mustang crate. Now that c-scape and c-gull have moved to their own
repos, this is no longer needed.

This PR combines the two, putting the mustang crate at the top level,
making the repository structure simpiler.